requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
celery>=5.4,<5.5
redis>=5.2,<5.3
psycopg2-binary>=2.9,<2.10
python-dotenv>=1.0,<1.1
uvicorn>=0.44,<0.45
